gr-dtv: Add code review updates.
