+++ gr-dtv/include/gnuradio/dtv/dvb_config.h
@@ -21,40 +21,6 @@
 #ifndef INCLUDED_DTV_DVB_CONFIG_H
 #define INCLUDED_DTV_DVB_CONFIG_H
 
-#define TRUE 1
-#define FALSE 0
-
-#define BB_HEADER_LENGTH_BITS 72
-
-// BB HEADER fields
-#define TS_GS_TRANSPORT 3
-#define TS_GS_GENERIC_PACKETIZED 0
-#define TS_GS_GENERIC_CONTINUOUS 1
-#define TS_GS_RESERVED 2
-
-#define SIS_MIS_SINGLE 1
-#define SIS_MIS_MULTIPLE 0
-
-#define CCM 1
-#define ACM 0
-
-#define ISSYI_ACTIVE 1
-#define ISSYI_NOT_ACTIVE 0
-
-#define NPD_ACTIVE 1
-#define NPD_NOT_ACTIVE 0
-
-#define FRAME_SIZE_NORMAL 64800
-#define FRAME_SIZE_SHORT  16200
-
-// BCH Code
-#define BCH_CODE_N8  0
-#define BCH_CODE_N10 1
-#define BCH_CODE_N12 2
-#define BCH_CODE_S12 3
-
-#define LDPC_ENCODE_TABLE_LENGTH (FRAME_SIZE_NORMAL * 10)
-
 namespace gr {
   namespace dtv {
     enum dvb_standard_t {
