+++ gr-dtv/lib/dvb/dvb_bbscrambler_bb_impl.cc
